+++ src/main/java/com/bookmark/library/view/MainMenuView.java
@@ -1,6 +1,5 @@
 package com.bookmark.library.view;
 
-import com.bookmark.library.model.Login.Main.LoginPage;
 import com.bookmark.library.util.IO;
 
 public class MainMenuView {
